task 15

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,12 +7,12 @@
 int main(void)
 {
 	int i;
-	unsigned long int next, first = 1, second = 2;
+	unsigned long long int next, first = 1, second = 2;
 
-	printf("%lu, ", first);
+	printf("%llu, ", first);
 	for (i = 1 ; i < 50 ; i++)
 	{
-		printf("%lu", second);
+		printf("%llu", second);
 		next = first + second;
 		first = second;
 		second = next;
